tend: extract agent_telegram helpers — router 607 → 167 lines; heal the large-module line back to 44
A FastAPI router carrying 25 pure formatting/URL/classification helpers alongside
its one webhook endpoint was poorly factored. Moved the helpers (+ their
constants) to a sibling module agent_telegram_format.py; the endpoint imports them
back. Behaviour-preserving (pure functions, verbatim sed-move) — verified by ruff
(clean), py_compile, import resolution, and a completeness check that every helper
the endpoint calls is imported back. Also removed a dead `task_url` assignment ruff
flagged in format_task_alert.

Measured against main (#2535): large_modules 45 → 44, risk 991 → 983 — the body
moves DOWN on both axes. This reverses the baseline bump PR #2534 made under
main-drift pressure (it raised large 44 → 45 instead of healing): the line goes
back to 44, earned by a real heal, not by raising the threshold.

Baseline refreshed to the healed reality (44 / 983). Risk sits above the prior 976
ONLY because #2535 introduced a >1000-line module (very_large 15 → 16) — that
module is the real next heal (the audit's modularization-review), named not
dismissed; this change pushed risk down 8 toward it, it did not ratchet up.
